+++ runtime/db/db.go
@@ -51,10 +51,6 @@ func Stop(ctx context.Context) {
 	close(globalRuntimePostgresWriter.quit)
 	<-globalRuntimePostgresWriter.done
 	pool.Close()
-
-	if _embeddedPostgresDB != nil {
-		shutdownEmbeddedPostgresDB(ctx)
-	}
 }
 
 func logDbWarningOrError(ctx context.Context, err error, msg string) {
@@ -401,13 +397,6 @@ func Initialize(ctx context.Context) {
 		return
 	}
 
-	if useEmbeddedPostgres() {
-		if err := prepareEmbeddedPostgresDB(ctx); err != nil {
-			logger.Fatal(ctx).Err(err).Msg("Failed to prepare embedded Postgres database.")
-			return
-		}
-	}
-
 	// this will initialize the pool and start the worker
 	_, err := globalRuntimePostgresWriter.GetPool(ctx)
 	if err != nil {
@@ -460,11 +449,7 @@ func useModusDB() bool {
 			}
 		}
 
-		if app.IsDevEnvironment() {
-			_useModusDB = !useEmbeddedPostgres()
-		} else {
-			_useModusDB = false
-		}
+		_useModusDB = app.IsDevEnvironment()
 	})
 	return _useModusDB
 }