+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # Change Log
 
+## UNRELEASED
+
+- feat: remove embedded PostgresDB [#870](https://github.com/hypermodeinc/modus/pull/870)
+
 ## 2025-05-29 - Runtime 0.18.0-alpha.4
 
 - chore: rename agent lifecycle methods and APIs [#858](https://github.com/hypermodeinc/modus/pull/858)